fix circular imports and class variable
