+++ src/neuroconv/tools/nwb_helpers/__init__.py
@@ -1,8 +1,6 @@
 """Collection of Pydantic models and helper functions for configuring dataset IO parameters for different backends."""
-from ._dataset_configuration import (
-    get_default_backend_configuration,
-    get_default_dataset_io_configurations,
-)
+from ._dataset_configuration import get_default_dataset_io_configurations
+from ._backend_configuration import get_default_backend_configuration
 from ._metadata_and_file_helpers import (
     add_device_from_metadata,
     get_default_nwbfile_metadata,
@@ -10,23 +8,20 @@ from ._metadata_and_file_helpers import (
     make_nwbfile_from_metadata,
     make_or_load_nwbfile,
 )
-from ._models._base_models import DatasetInfo
-from ._models._hdf5_models import (
-    AVAILABLE_HDF5_COMPRESSION_METHODS,
-    HDF5BackendConfiguration,
-    HDF5DatasetIOConfiguration,
-)
-from ._models._zarr_models import (
-    AVAILABLE_ZARR_COMPRESSION_METHODS,
-    ZarrBackendConfiguration,
-    ZarrDatasetIOConfiguration,
-)
+from ._configuration_models._base_dataset_io import DatasetInfo
+from ._configuration_models._hdf5_dataset_io import AVAILABLE_HDF5_COMPRESSION_METHODS, HDF5DatasetIOConfiguration
+from ._configuration_models._hdf5_backend import HDF5BackendConfiguration
+from ._configuration_models._zarr_dataset_io import AVAILABLE_ZARR_COMPRESSION_METHODS, ZarrDatasetIOConfiguration
+from ._configuration_models._zarr_backend import ZarrBackendConfiguration
 
 BACKEND_CONFIGURATIONS = dict(hdf5=HDF5BackendConfiguration, zarr=ZarrBackendConfiguration)
 DATASET_IO_CONFIGURATIONS = dict(hdf5=HDF5DatasetIOConfiguration, zarr=ZarrDatasetIOConfiguration)
 
 __all__ = [
+    "AVAILABLE_HDF5_COMPRESSION_METHODS",
+    "AVAILABLE_ZARR_COMPRESSION_METHODS",
     "get_default_dataset_io_configurations",
+    "get_default_backend_configuration",
     "add_device_from_metadata",
     "get_default_nwbfile_metadata",
     "get_module",
